fixed bugs

+++ public/src/js/api-clients/face-client.js
@@ -4,15 +4,19 @@ $('#cancel-btnFace').on('click',function(event){
 	event.preventDefault();	
 	var username = $('#username').val().trim();
 	window.location.href = "#/";
-	
+	var apiPhase = $("#apiPhase").text();
 
-	$.ajax({
-		method: 'DELETE',
-		url: '/api/delete/db/' + username
-	}).done(function(res){
-		console.log("DB delete at face step: success")
-			
-	})
+	if(apiPhase == "signup"){
+		$.ajax({
+			method: 'DELETE',
+			url: '/api/delete/db/' + username
+		}).done(function(res){
+			console.log("DB delete at face step: success")
+				
+		});
+	} else if(apiPhase == "login"){
+		window.location.replace(window.location.origin+"/login/face/"+username);
+	}
 
 });
